refactor(std): move timers + time to rts-std
timers (global setTimeout/setInterval/queueMicrotask drain) -> rts-std/src/globals/timers; time (rts:time ns) -> rts-std/src/time. Desbloqueados pelo move de text_encoding (timers chama drain_microtasks, agora intra-std).

Refs: gc::handles -> rts_engine::heap::handles; globals::text_encoding/timers -> crate::globals::*. Facade no runtime via pub use.

Gate: build limpo; rts.exe test 1710/1710; smoke performance.now JIT==AOT OK. setTimeout (macrotask) só dispara em JIT (AOT sem event loop drain, pré-existente).
